+++ include/battery_utils.h
@@ -6,7 +6,6 @@
 
 namespace BATTERY_Utils {
 
-    String  generateEncodedTelemetry(float voltage);
     String  getPercentVoltageBattery(float voltage);
     void    checkVoltageWithoutGPSFix();
